The IP ID field is not known.  Although, the fragment offset is generaly known, it comes after the IP ID.  It is not useful since the caller to known_clear expects a full known keystream and not a keystream with some incorrect bytes in it.
Perhaps we can modify the PTW attack to ignore the keystream's IP ID bytes and
to not calculate the respective bytes of the key when using those keystreams.
That is, use IP packets to calculate only key bytes 0--9, and if possible,
12--13, but not 10--11 which will be derived from the (unknown) IP ID.  Those
bytes can be attacked with ARP as usual.  Also note, that 40-bit keys only
require 8 bytes of keystream which is known in IP [LLC/SNAP].  [This discussion
asusmes that X keystream bytes are necessary to decrypt a key of length X-3.]


git-svn-id: http://svn.aircrack-ng.org/branch/0.9-dev@286 28c6078b-6c39-48e3-add9-af49d547ecab