Airodump-ng: Automatically colorize airodump.
I've patched airodump so that upon a key press it automatically adds colors only to the APs that have associated clients. Also clients that are not associated are greyed out. This makes it much easier to see who's-connected-to-who and who-is-probing-for-who.

Press o to turn the color on, and p to turn it off.