Fix wpaclean to write all, and not only the last captured handshake to the output file.
In save_network(), open_pcap() is called for each detected network
in the input capture files. open_pcap() truncates the existing
file, and starts from zero with each network it finds.
So only call open_pcap() on the first network/handshake it detects,
so that subsequtent handshakes are added to the output capture file.