fix(airdrop-ng,airgraph-ng): regex the ESSID for CAPR
This adds the ability to extract from a csv dump ESSID containing
commas when creating a CAPR image with airgraph-ng.

When in a csv-comma-separeted file an ESSID containing a comma is
truncated when the entry is comma-split.

ESSID: "this is, my ap"

     ID-length     ESSID       Key
e.g. "..., 14, this is, my ap, "

                 LAN IP         ESSID
e.g. "...,   0.  0.  0.  0, this is, my ap"

A working approach to get the whole ESSID is to find its position
in the entry, and then to use ID-length (the ESSID's length) when
it's available to grab the ESSID, otherwise the ESSID is possibly
after LAN IP.