fix(airdrop-ng,airgraph-ng,airodump-ng): regex the ESSIDs for CPG
This adds the ability to extract from a csv dump ESSID containing
commas when creating a CPG image with airgraph-ng.

When in a csv-comma-separeted file an ESSID containing a comma is
truncated when the entry is comma-split.

With this commit, dump_write.c (aka airodump-ng) now will precede
each probed ESSID with its own length when writing the csv file.

ESSID: "this is, my ap", "second ap"

                 Probed ESSIDs
e.g. "...,14,this is, my ap,9,second ap"

A working approach to get the whole ESSID is to find the position
of the Probed ESSIDs in the entry, and then to grab the ESSID via
its provided length.