build: add option to disable bsd library inclusion
It might be needed to disable bsd inclusion and fallback to the compat
functions even if bsd headers are detected.

This is the case when multiple library are cross-compiled and someone
wants to explicitly compile aircrack-ng without linking to bsd library.

With the current implementation, if a bsd header is detected, the bsd
library is always linked even if unwanted. Add option to configure this
with the combo --with-libbsd=yes|no|auto with auto set by default.

Also add an extra featurw with introducing the possibility of requiring
the bsd library and fail the configure phase.

Signed-off-by: Christian Marangi <ansuelsmth@gmail.com>